requirements.txt
pillar-youtube-upload